fix(world): w is a coordinate you travel along, not a menu of worlds
The previous version was wrong about what a fourth dimension is, and it
was wrong in the way that matters: pressing a key jumped to the next
integer slice, so w was a selector. You teleported between discrete
worlds instead of moving through one. It did not read as four-dimensional
because it was not - it was a level swap with a nice transition.

Three changes make it an axis.

w is a float. TerrainGen4D::height_at and fill_chunk take a continuous w
rather than an integer slice index. An integer w makes the fourth axis a
menu by construction; nothing downstream can recover from that.

Travel is held, not pressed. `,` and `.` use key_down, so holding one
slides the player through w the way holding W slides them through z, and
the terrain morphs continuously while held. Shift sprints, so the fourth
axis handles like the other three.

Rebuilding is decoupled from moving. Travel is continuous; re-meshing is
not affordable per frame, since moving along w changes 100% of the chunks
in the window. So the world is rebuilt only once the player has drifted
kSliceRemeshStep (0.12) from the w the geometry was built at, and the HUD
shows both numbers - the lag between them is the design, not a defect.

Speeds are bounded by the worker pool rather than by feel. A rebuild
re-requests every chunk in the window (625 at radius 12) and the pool
sustains ~2,200 chunks/sec, so a rebuild costs ~0.28 s there; at a 0.12
threshold that allows about 0.43 units/sec before the world stops keeping
up. Walk is 0.4. Sprint is 1.2 and does outrun the pool at radius 12,
which is the honest trade rather than a bug: hold shift and the terrain
lags behind you.

Rebuilds now go nearest-to-the-player first. In scan order the world
changed from one corner outward, which reads as a glitch; nearest-first
morphs what you are looking at immediately and lets the far edge catch up
over the following frames. Ties break on (z, x) so the order stays
deterministic.

One world unit of w changes about half the columns, so walking crosses a
visibly different world every two and a half seconds - a pace for
exploring an axis rather than flicking through it.

Unmoved: --validate reports gpu_mesh_mb=10.99, six ctest suites pass,
--verify-4d still shows changed=1 returned=1 with clean meshes at every
stop.

+++ src/world/terrain_gen4d.cpp
@@ -26,10 +26,10 @@ TerrainGen4D::TerrainGen4D(std::uint32_t seed)
       temp_(seed + 5),
       cave_a_(seed + 6), cave_b_(seed + 7) {}
 
-int TerrainGen4D::height_at(int wx, int wz, int w) const {
+int TerrainGen4D::height_at(int wx, int wz, float w) const {
     const float x = static_cast<float>(wx);
     const float z = static_cast<float>(wz);
-    const float fw = static_cast<float>(w) * kWScale;
+    const float fw = w * kWScale;
 
     // Domain warp, then three octave stacks over a 3D slice of the 4D
     // field: world x, world z and w, with the noise's remaining axis
@@ -58,10 +58,10 @@ int TerrainGen4D::height_at(int wx, int wz, int w) const {
         1, kChunkSizeY - 1);
 }
 
-void TerrainGen4D::fill_chunk(int chunk_x, int chunk_z, int w, Chunk& out) const {
+void TerrainGen4D::fill_chunk(int chunk_x, int chunk_z, float w, Chunk& out) const {
     const int origin_x = chunk_x * kChunkSizeX;
     const int origin_z = chunk_z * kChunkSizeZ;
-    const float fw = static_cast<float>(w) * kWScale;
+    const float fw = w * kWScale;
 
     int  surface[kChunkSizeZ][kChunkSizeX];
     bool is_desert[kChunkSizeZ][kChunkSizeX];
